refactor common folder
