+++ app/api.ts
@@ -1,5 +1,5 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
-import { PARTNER_ACCESS_CODE_STATUS } from '../common/constants';
+import { PARTNER_ACCESS_CODE_STATUS } from '../constants/responses';
 import { PartnerAccess } from './partnerAccessSlice';
 import { Partner } from './partnerSlice';
 import { RootState } from './store';
