+++ constants/languages.ts
@@ -1,0 +1,4 @@
+export enum LANGUAGES {
+  en = 'en',
+  es = 'es',
+}
